extract index page generation
