+++ src/lib.rs
@@ -1,0 +1,2 @@
+mod page;
+pub use crate::page::genpage;
